Lua Cleanup 2
Merged libaries of minimizer and checklibary for #37
